requirements.txt
Django==2.2
django-cors-headers==2.5.2
django-countries==5.3.3
django-phonenumber-field==2.3.1
phonenumbers==8.10.8
django-money==0.14.4
django-redis
django-inline-actions==2.1.0
django-admin-rangefilter==0.3.15
django-grappelli==2.12.2
psycopg2-binary==2.8.1
Pillow==6.0.0